[fix] refs #86 ログインループの不具合を解消
