+++ src/.env
@@ -1,1 +1,1 @@
-VUE_APP_GRAPHQL_HTTP=http://localhost:8000/gql/
+VUE_APP_GRAPHQL_HTTP=https://montage-graphql-api.herokuapp.com/gql/
