+++ src/package.json
@@ -10,6 +10,7 @@
     "test:unit": "vue-cli-service test:unit"
   },
   "dependencies": {
+    "@types/auth0-js": "^9.10.6",
     "@types/graphql": "^14.5.0",
     "@types/semantic-ui": "^2.2.7",
     "apollo-cache-inmemory": "^1.6.3",
