+++ src/src/main.ts
@@ -28,8 +28,8 @@ const errorLink = onError(({ graphQLErrors, networkError }) => {
 
 // HTTP connexion to the API
 const httpLink = new HttpLink({
-  // uri: 'https://montage-graphql-api.herokuapp.com/gql/',
-  uri: 'localhost:8000/gql/',
+  uri: 'https://montage-graphql-api.herokuapp.com/gql/',
+  // uri: 'localhost:8000/gql/',
 });
 
 // Cache implementation
