+++ src/src/shims-vue.d.ts
@@ -10,4 +10,3 @@ declare module 'vue/types/vue' {
     $auth: any;
   }
 }
-
