+++ src/tslint.json
@@ -4,7 +4,7 @@
         "tslint:recommended"
     ],
     "linterOptions": {
-      "exclude": ["./node_modules/**", "./src/**/*.js", "./src/*.js"]
+      "exclude": ["./node_modules/**", "./src/**/*.js", "./src/*.js", "*.json"]
     },
     "jsRules": {},
     "rules": {
